devops8/restfuldemo/requirement.txt
Django==1.11.13
djangorestframework
django-filter==1.1.0
Markdown==2.6.11
mysqlclient
pytz==2018.4
requests
ansible
djangorestframework-jwt

tencentcloud-sdk-python